Time: 397 ms (35.92%), Space: 100.6 MB (94.10%) - LeetHub

+++ 2171-removing-minimum-number-of-magic-beans/2171-removing-minimum-number-of-magic-beans.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long minimumRemoval(vector<int>& beans) {
-     long long n=beans.size();
+     int n=beans.size();
      sort(beans.begin(),beans.end());
      long long total=0,prev=0;
      for(int i=0;i<n;i++)
